+++ packages/lightspark-sdk/src/objects/GraphNode.ts
@@ -4,13 +4,14 @@ import { type Query } from "@lightsparkdev/core";
 import autoBind from "auto-bind";
 import type LightsparkClient from "../client.js";
 import BitcoinNetwork from "./BitcoinNetwork.js";
+import type Entity from "./Entity.js";
 import type Node from "./Node.js";
 import type NodeAddressType from "./NodeAddressType.js";
 import type NodeToAddressesConnection from "./NodeToAddressesConnection.js";
 import { NodeToAddressesConnectionFromJson } from "./NodeToAddressesConnection.js";
 
 /** This object represents a node that exists on the Lightning Network, including nodes not managed by Lightspark. You can retrieve this object to get publicly available information about any node on the Lightning Network. **/
-class GraphNode implements Node {
+class GraphNode implements Node, Entity {
   constructor(
     public readonly id: string,
     public readonly createdAt: string,
@@ -18,10 +19,10 @@ class GraphNode implements Node {
     public readonly bitcoinNetwork: BitcoinNetwork,
     public readonly displayName: string,
     public readonly typename: string,
-    public readonly alias?: string,
-    public readonly color?: string,
-    public readonly conductivity?: number,
-    public readonly publicKey?: string,
+    public readonly alias?: string | undefined,
+    public readonly color?: string | undefined,
+    public readonly conductivity?: number | undefined,
+    public readonly publicKey?: string | undefined,
   ) {
     autoBind(this);
   }
@@ -72,6 +73,21 @@ ${FRAGMENT}
 `,
       variables: { id },
       constructObject: (data: any) => GraphNodeFromJson(data.entity),
+    };
+  }
+
+  public toJson() {
+    return {
+      __typename: "GraphNode",
+      graph_node_id: this.id,
+      graph_node_created_at: this.createdAt,
+      graph_node_updated_at: this.updatedAt,
+      graph_node_alias: this.alias,
+      graph_node_bitcoin_network: this.bitcoinNetwork,
+      graph_node_color: this.color,
+      graph_node_conductivity: this.conductivity,
+      graph_node_display_name: this.displayName,
+      graph_node_public_key: this.publicKey,
     };
   }
 }