+++ packages/lightspark-sdk/src/objects/ReleaseChannelPerCommitmentSecretInput.ts
@@ -1,6 +1,6 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
 
-type ReleaseChannelPerCommitmentSecretInput = {
+interface ReleaseChannelPerCommitmentSecretInput {
   /** The unique identifier of the channel. **/
   channelId: string;
 
@@ -9,7 +9,7 @@ type ReleaseChannelPerCommitmentSecretInput = {
 
   /** The index associated with the per-commitment secret. **/
   perCommitmentIndex: number;
-};
+}
 
 export const ReleaseChannelPerCommitmentSecretInputFromJson = (
   obj: any,
@@ -21,6 +21,17 @@ export const ReleaseChannelPerCommitmentSecretInputFromJson = (
     perCommitmentIndex:
       obj["release_channel_per_commitment_secret_input_per_commitment_index"],
   } as ReleaseChannelPerCommitmentSecretInput;
+};
+export const ReleaseChannelPerCommitmentSecretInputToJson = (
+  obj: ReleaseChannelPerCommitmentSecretInput,
+): any => {
+  return {
+    release_channel_per_commitment_secret_input_channel_id: obj.channelId,
+    release_channel_per_commitment_secret_input_per_commitment_secret:
+      obj.perCommitmentSecret,
+    release_channel_per_commitment_secret_input_per_commitment_index:
+      obj.perCommitmentIndex,
+  };
 };
 
 export default ReleaseChannelPerCommitmentSecretInput;