+++ packages/lightspark-sdk/src/objects/SetInvoicePaymentHashOutput.ts
@@ -1,8 +1,8 @@
 // Copyright ©, 2023-present, Lightspark Group, Inc. - All Rights Reserved
 
-type SetInvoicePaymentHashOutput = {
+interface SetInvoicePaymentHashOutput {
   invoiceId: string;
-};
+}
 
 export const SetInvoicePaymentHashOutputFromJson = (
   obj: any,
@@ -10,6 +10,13 @@ export const SetInvoicePaymentHashOutputFromJson = (
   return {
     invoiceId: obj["set_invoice_payment_hash_output_invoice"].id,
   } as SetInvoicePaymentHashOutput;
+};
+export const SetInvoicePaymentHashOutputToJson = (
+  obj: SetInvoicePaymentHashOutput,
+): any => {
+  return {
+    set_invoice_payment_hash_output_invoice: { id: obj.invoiceId },
+  };
 };
 
 export const FRAGMENT = `