+++ packages/wallet-sdk/src/objects/TerminateWalletOutput.ts
@@ -3,9 +3,9 @@
 import type Wallet from "./Wallet.js";
 import { WalletFromJson } from "./Wallet.js";
 
-type TerminateWalletOutput = {
+interface TerminateWalletOutput {
   wallet: Wallet;
-};
+}
 
 export const TerminateWalletOutputFromJson = (
   obj: any,
@@ -13,6 +13,13 @@ export const TerminateWalletOutputFromJson = (
   return {
     wallet: WalletFromJson(obj["terminate_wallet_output_wallet"]),
   } as TerminateWalletOutput;
+};
+export const TerminateWalletOutputToJson = (
+  obj: TerminateWalletOutput,
+): any => {
+  return {
+    terminate_wallet_output_wallet: obj.wallet.toJson(),
+  };
 };
 
 export const FRAGMENT = `