+++ apps/examples/remote-signing-server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @lightsparkdev/remote-signing-server
 
+## 0.0.10
+
+### Patch Changes
+
+- Updated dependencies [4857f66]
+  - @lightsparkdev/lightspark-sdk@1.2.0
+
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [c6fceaa]
+  - @lightsparkdev/lightspark-sdk@1.1.7
+
 ## 0.0.8
 
 ### Patch Changes