+++ apps/examples/streaming-wallet-extension/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @lightsparkdev/streaming-wallet-extension
 
+## 1.0.39
+
+### Patch Changes
+
+- Updated dependencies [4857f66]
+  - @lightsparkdev/lightspark-sdk@1.2.0
+
+## 1.0.38
+
+### Patch Changes
+
+- Updated dependencies [c6fceaa]
+  - @lightsparkdev/lightspark-sdk@1.1.7
+
 ## 1.0.37
 
 ### Patch Changes